+++ crates/oasis7/src/runtime/world/mod.rs
@@ -28,6 +28,7 @@ mod policy;
 mod release_manifest;
 mod resources;
 mod restricted_claim_grants;
+mod rollback;
 mod rules;
 mod scheduling;
 mod snapshot;
